+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvDataSource.java
@@ -8,6 +8,7 @@ package edu.ie3.datamodel.io.source.csv;
 import edu.ie3.datamodel.exceptions.SourceException;
 import edu.ie3.datamodel.io.connectors.CsvFileConnector;
 import edu.ie3.datamodel.io.factory.EntityFactory;
+import edu.ie3.datamodel.io.factory.SimpleEntityData;
 import edu.ie3.datamodel.io.factory.input.AssetInputEntityData;
 import edu.ie3.datamodel.io.factory.input.NodeAssetInputEntityData;
 import edu.ie3.datamodel.io.naming.EntityPersistenceNamingStrategy;
@@ -16,6 +17,7 @@ import edu.ie3.datamodel.models.input.AssetInput;
 import edu.ie3.datamodel.models.input.AssetTypeInput;
 import edu.ie3.datamodel.models.input.NodeInput;
 import edu.ie3.datamodel.models.input.OperatorInput;
+import edu.ie3.datamodel.models.result.ResultEntity;
 import edu.ie3.datamodel.utils.validation.ValidationUtils;
 import edu.ie3.util.StringUtils;
 import java.io.BufferedReader;
@@ -619,5 +621,19 @@ public abstract class CsvDataSource {
       Collection<OperatorInput> operators) {
     return nodeAssetInputEntityDataStream(assetInputEntityDataStream(entityClass, operators), nodes)
         .map(dataOpt -> dataOpt.flatMap(factory::get));
+  }
+
+  /**
+   * Returns a stream of {@link SimpleEntityData} for result entity classes, using a
+   * fields-to-attributes map.
+   *
+   * @param entityClass the entity class that should be build
+   * @param <T> Type of the {@link ResultEntity} to expect
+   * @return stream of {@link SimpleEntityData}
+   */
+  protected <T extends ResultEntity> Stream<SimpleEntityData> simpleEntityDataStream(
+      Class<T> entityClass) {
+    return buildStreamWithFieldsToAttributesMap(entityClass, connector)
+        .map(fieldsToAttributes -> new SimpleEntityData(fieldsToAttributes, entityClass));
   }
 }